Refactored types based on feedback
