+++ plugins/search-backend-module-pg/src/PgSearchEngine/index.ts
@@ -14,7 +14,13 @@
  * limitations under the License.
  */
 export { PgSearchEngine } from './PgSearchEngine';
-export type { ConcretePgSearchQuery } from './PgSearchEngine';
+export type {
+  ConcretePgSearchQuery,
+  PgSearchQueryTranslatorOptions,
+  PgSearchQueryTranslator,
+  PgSearchOptions,
+  PgSearchHighlightOptions,
+} from './PgSearchEngine';
 export type {
   PgSearchEngineIndexer,
   PgSearchEngineIndexerOptions,
