--- plugins/search-backend-module-pg/src/types.ts
@@ -1,43 +1,0 @@
-/*
- * Copyright 2022 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { PluginDatabaseManager } from '@backstage/backend-common';
-import { Config } from '@backstage/config';
-
-/**
- * Options to instantiate PgSearchEngine
- * @public
- */
-export type PgSearchOptions = {
-  config: Config;
-  database: PluginDatabaseManager;
-};
-
-/**
- * Options for highlighting search terms
- * @public
- */
-export type PgSearchHighlightOptions = {
-  useHighlight?: boolean;
-  maxWords?: number;
-  minWords?: number;
-  shortWord?: number;
-  highlightAll?: boolean;
-  maxFragments?: number;
-  fragmentDelimiter?: string;
-  preTag: string;
-  postTag: string;
-};
